requirements.txt
Flask==3.0.0
Flask-RESTX==1.3.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv
